+++ Assets/Scenes/Environment/JUMPZONE.cs
@@ -1,0 +1,32 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class JUMPZONE : MonoBehaviour
+{
+    // Start is called before the first frame update
+
+    
+
+
+    void Start()
+    {
+       
+    }
+
+    // Update is called once per frame
+    void Update()
+    {
+
+    }
+    private void OnCollisionEnter2D(Collision2D collision)
+    {
+        if (collision.gameObject.tag == "Player")
+        {
+            collision.gameObject.GetComponent<movement>().JumpZone();
+
+
+        }
+    }
+   
+}
