+++ Assets/Scenes/Environment/JUMPZONE.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: a711b886cfa7cef4986ca89ed2ca9204
+guid: f7ede69c8a19f4148a1ce78fbdfa6955
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
