+++ Assets/Scenes/Environment/Speedzone.cs
@@ -2,31 +2,34 @@ using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 
-public class Speedzone : MonoBehaviour
+public class SPEEDZONE : MonoBehaviour
 {
     // Start is called before the first frame update
-
-    
-
-
     void Start()
     {
-       
+        
     }
 
     // Update is called once per frame
     void Update()
     {
-
+        
     }
-    private void OnCollisionEnter2D(Collision2D collision)
+    void OnCollisionEnter2D(Collision2D collision)
     {
         if (collision.gameObject.tag == "Player")
         {
-            collision.gameObject.GetComponent<movement>().JumpZone();
+            collision.gameObject.GetComponent<movement>().Speedzone();
+        }
+    }
+    private void OnCollisionExit2D(Collision2D collision)
+    {
+        if (collision.gameObject.tag == "Player")
+        {
+            collision.gameObject.GetComponent<movement>().Originalspeed();
 
 
         }
     }
-   
+
 }
