+++ UserSettings/EditorUserSettings.asset
@@ -14,6 +14,9 @@ EditorUserSettings:
     RecentlyUsedScenePath-2:
       value: 22424703114646680e0b0227036c6c111b07142f1f2b233e28694c7df7ee3d2cfb
       flags: 0
+    RecentlyUsedScenePath-3:
+      value: 22424703114646680e0b0227036c7a1e001e0a252225233e39662e32eff0383dd1e433e7e27a6e722c0ce6281d
+      flags: 0
     vcSharedLogLevel:
       value: 0d5e400f0650
       flags: 0
